make a relation between post and user where we see user proprety in post

+++ app/routers/post.py
@@ -26,7 +26,7 @@ async def get_posts(db: db_dependency, current_user: int = Depends(oauth2.get_cu
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.Post)
 async def create_post(post: schemas.PostCreate, db: db_dependency, current_user: int = Depends(oauth2.get_current_user)):
 
-    new_post = models.Post(**post.dict())
+    new_post = models.Post(owner_id=current_user.id, **post.dict())
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
@@ -63,6 +63,12 @@ async def delete_post(id: int, db: db_dependency, current_user: int = Depends(oa
             detail=f"post with id: {id} was not found"
         )
 
+    if post.owner_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail=f"Not authorized to perform requested action"
+        )
+
     post_query.delete(synchronize_session=False)
     db.commit()
 
@@ -81,6 +87,12 @@ async def update_post(id: int, updated_post: schemas.PostCreate, db: db_dependen
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"post with id: {id} was not found"
+        )
+    
+    if post.owner_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail=f"Not authorized to perform requested action"
         )
 
     post_query.update(updated_post.dict(), synchronize_session=False)
